requirements.txt
numpy==1.21.0
pandas==1.2.5
google-cloud-aiplatform==1.14.0
matplotlib==3.4.2
opencv-python-headless